routes: implement story delete and vote
Implement story delete.  It is missing a button in the UI.  It has not
been tested!

Implement question vote and add forms around the voting buttons.  It
has been minimally tested.

Moved routes/feed.js and routes/story.js inside of routes.js for easy
access to rendering the feed and a story view with custom error
messages.  We should probably clean that code up at some point, but it
seems to work well for now.

There was an error in the backend API spec where question vote was
specified to return a question, but it shouldn't.