confirm passwords locally
There are two password confirmations in the app:
 - At account creation
 - At account password change

For the former, the only password validation was from a Javascript file,
meaning if a client didn't have Javascript enabled there would be
absolutely no checking to verify that the users's passwords fields
matched.  In the latter, the password equality was checked on the
backend.

This change adds a check for the former and the latter in the frontend
app.

With this, the backend should never receive inequal passwords for
password change from this HTML frontend anymore.